Complete reorganization of schemas, splitting into separate files
