+++ api/routes/old/alerts.py
@@ -18,7 +18,7 @@ from typing import List
 from fastapi import APIRouter, Depends, Security, HTTPException
 from sqlalchemy.orm import Session
 
-from api import schemas
+from api.schemas import meter_schemas
 from api.models import Alert
 from api.route_util import _add, _patch
 from api.security import get_current_user, scoped_user
@@ -29,12 +29,12 @@ alert_router = APIRouter()
 write_user = scoped_user(["read", "alerts:write"])
 
 
-@alert_router.get("/alerts", response_model=List[schemas.Alert], tags=["alerts"])
+@alert_router.get("/alerts", response_model=List[meter_schemas.Alert], tags=["alerts"])
 async def read_alerts(db: Session = Depends(get_db)):
     return db.query(Alert).all()
 
 
-@alert_router.get("/alerts/{alert_id}", response_model=schemas.Alert, tags=["alerts"])
+@alert_router.get("/alerts/{alert_id}", response_model=meter_schemas.Alert, tags=["alerts"])
 async def read_alerts(alert_id: int, db: Session = Depends(get_db)):
     return db.query(Alert).filter(Alert.id == alert_id).first()
 
@@ -42,21 +42,21 @@ async def read_alerts(alert_id: int, db: Session = Depends(get_db)):
 @alert_router.post(
     "/alerts",
     dependencies=[Depends(write_user)],
-    response_model=schemas.Alert,
+    response_model=meter_schemas.Alert,
     tags=["alerts"],
 )
-async def add_alerts(alert: schemas.AlertCreate, db: Session = Depends(get_db)):
+async def add_alerts(alert: meter_schemas.AlertCreate, db: Session = Depends(get_db)):
     return _add(db, Alert, alert)
 
 
 @alert_router.patch(
     "/alerts/{alert_id}",
     dependencies=[Depends(write_user)],
-    response_model=schemas.Alert,
+    response_model=meter_schemas.Alert,
     tags=["alerts"],
 )
 async def patch_alerts(
-    alert_id: int, obj: schemas.AlertPatch, db: Session = Depends(get_db)
+    alert_id: int, obj: meter_schemas.AlertPatch, db: Session = Depends(get_db)
 ):
     return _patch(db, Alert, alert_id, obj)
